+++ test/dynamic_event_store_test.exs
@@ -1,4 +1,4 @@
-defmodule DynamicEventStoreTest do
+defmodule EventStore.DynamicEventStoreTest do
   use EventStore.StorageCase
 
   alias EventStore.EventFactory