+++ test/multi_event_store_test.exs
@@ -1,4 +1,4 @@
-defmodule MultiEventStoreTest do
+defmodule EventStore.MultiEventStoreTest do
   use EventStore.StorageCase
 
   alias EventStore.{Config, EventData, EventFactory, RecordedEvent, Storage}