+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Remove distributed registry ([#210](https://github.com/commanded/eventstore/pull/210)).
 - Hibernate subscription process after inactivity ([#214](https://github.com/commanded/eventstore/pull/214)).
 - Runtime event store configuration ((#217)[https://github.com/commanded/eventstore/pull/217]).
+- Shared database connection pools ([#216](https://github.com/commanded/eventstore/pull/216)).
 
 ### Bug fixes